tests/wpt/web-platform-tests/css/css-text/parsing/word-boundary-expansion-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text: getComputedStyle().wordBoundaryExpansion</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-boundary-expansion">
<meta name="assert" content="word-boundary-expansion computed value is specified keyword.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("word-boundary-expansion", "none");
test_computed_value("word-boundary-expansion", "space");
test_computed_value("word-boundary-expansion", "ideographic-space");
</script>
</body>
</html>
